+++ src/main/java/net/voidz/VoidMain.java
@@ -4,6 +4,7 @@ import net.fabricmc.api.ModInitializer;
 import net.voidz.init.BlockInit;
 import net.voidz.init.ConfigInit;
 import net.voidz.init.DimensionInit;
+import net.voidz.init.EventInit;
 import net.voidz.init.ItemInit;
 
 public class VoidMain implements ModInitializer {
@@ -14,6 +15,7 @@ public class VoidMain implements ModInitializer {
         DimensionInit.init();
         ItemInit.init();
         BlockInit.init();
+        EventInit.init();
     }
 }